Desperados — Update script to 2.10 syntax

+++ play.it-2/games/play-desperados.sh
@@ -31,18 +31,16 @@ set -o errexit
 
 ###
 # Desperados: Wanted Dead or Alive
-# build native Linux packages from the original installers
+# build native packages from the original installers
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20181029.1
 
 # Set game-specific variables
 
 GAME_ID='desperados'
 GAME_NAME='Desperados: Wanted Dead or Alive'
-
-ARCHIVES_LIST='ARCHIVE_GOG'
 
 ARCHIVE_GOG='setup_desperados_wanted_dead_or_alive_2.0.0.6.exe'
 ARCHIVE_GOG_URL='https://www.gog.com/game/desperados_wanted_dead_or_alive'
@@ -50,69 +48,71 @@ ARCHIVE_GOG_MD5='8e2f4e2ade9e641fdd35a9dd36d55d00'
 ARCHIVE_GOG_VERSION='1.01-gog2.0.0.6'
 ARCHIVE_GOG_SIZE='810000'
 
-ARCHIVE_DOC1_PATH='app'
-ARCHIVE_DOC1_FILES='./manual.pdf ./readme.txt'
+ARCHIVE_DOC0_DATA_PATH='app'
+ARCHIVE_DOC0_DATA_FILES='manual.pdf readme.txt'
 
-ARCHIVE_DOC2_PATH='tmp'
-ARCHIVE_DOC2_FILES='./gog_eula.txt'
+ARCHIVE_DOC1_DATA_PATH='tmp'
+ARCHIVE_DOC1_DATA_FILES='gog_eula.txt'
 
 ARCHIVE_GAME_BIN_PATH='app/game'
-ARCHIVE_GAME_BIN_FILES='./*.dll ./*.exe'
+ARCHIVE_GAME_BIN_FILES='*.dll *.exe'
 
 ARCHIVE_GAME_DATA_PATH='app/game'
-ARCHIVE_GAME_DATA_FILES='./data'
+ARCHIVE_GAME_DATA_FILES='data'
 
 CONFIG_DIRS='./data/configuration'
 DATA_DIRS='./data/savegame ./data/levels/briefing/Restart'
 
 APP_MAIN_TYPE='wine'
 APP_MAIN_EXE='game.exe'
-APP_MAIN_ICON='./game.exe'
-APP_MAIN_ICON_RES='16 32'
+APP_MAIN_ICON='game.exe'
 
-PACKAGES_LIST='PKG_DATA PKG_BIN'
+PACKAGES_LIST='PKG_BIN PKG_DATA'
 
 PKG_DATA_ID="${GAME_ID}-data"
 PKG_DATA_DESCRIPTION='data'
 
 PKG_BIN_ARCH='32'
-PKG_BIN_DEPS_DEB="$PKG_DATA_ID, wine32-development | wine32 | wine-bin | wine-i386 | wine-staging-i386, wine:amd64 | wine"
-PKG_BIN_DEPS_ARCH="$PKG_DATA_ID wine"
+PKG_BIN_DEPS="$PKG_DATA_ID wine"
 
 # Load common functions
 
-target_version='2.0'
+target_version='2.10'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: ${XDG_DATA_HOME:="$HOME/.local/share"}
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
+fi
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
 fi
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
-
-PKG='PKG_BIN'
-organize_data 'GAME_BIN' "$PATH_GAME"
-
-PKG='PKG_DATA'
-organize_data 'DOC'       "$PATH_DOC"
-organize_data 'GAME_DATA' "$PATH_GAME"
-
-PKG='PKG_BIN'
-extract_and_sort_icons_from 'APP_MAIN'
-move_icons_to 'PKG_DATA'
-
+prepare_package_layout
 rm --recursive "$PLAYIT_WORKDIR/gamedata"
+
+# Extract icons
+
+PKG='PKG_BIN'
+icons_get_from_package 'APP_MAIN'
+icons_move_to 'PKG_DATA'
 
 # Write launchers
